Implemented save_file_as. Made several improvements to ric_value, it will not crash if ricword is NULL, you can change it after initialization and a few bugfixes. Improved the ricModel update, as reset will also reset the view which isn't wanted it it is just a single value that changed.
Fixed a some critical bugs in the write file procedure and centralized it. nxtVarRicWord is now also no longer completely inline.

All this is actually just bug-fixes to make the new addition work, a control to edit the Options opcode. It should be easy to extend for the other opcodes.

git-svn-id: svn://svn.code.sf.net/p/riccreator/code@44 dab27aee-19fe-a441-870f-52ce2490c35b